Move ApertureMask to separate module
